+++ main.cpp
@@ -3,7 +3,7 @@
 using namespace std;
 
 int main(void){
-    int num1, num2;
+    double num1, num2;
     string op;
     
     cout << "Please enter the operation you wish to use, followed by the first and second operands." << endl;
@@ -19,7 +19,7 @@ int main(void){
     return 0;
 }
 
-int calculator(int num1,int num2, string op){
+int calculator(double num1,double num2, string op){
     if(op == "+"){
         return num1+num2;
     }
